fix: create separate files for different functions
